refactored stacker tests
